feat: Read start banner info from a config file
rather than defining it directly inside the `.typ` sources. This
hopefully makes the config more accessible to people not familiar with
the syntax of typst. It also ensures that creating the banners doesn't
require modifications in the sources themselves. In addition, this
allows scripting the banner creation and integrating it into external
workflows as the config is easily created by other tools.