build(make): Pick up changes in the `config.yml`
and other source files needed to create the start and end banners. The
way the makefile was written previously, updates in these dependencies
were indeed recognized, but `make` wasn't compelled to rebuild the
actual outputs since their intermediates didn't exist and make saw no
reason to change that. By introducing a blank rule that only touches the
`.typ` sources (which, in turn, depend on the config etc.), the
intermediates are rebuilt correctly and the outputs are regenerated when
things change.