feat: Prepare the "Creative Commons" icons in typst
and replace the colors there rather than in the Makefile. On the one
hand this reduces the amount of utility files that must be written to
disk, reducing pollution of the repo with "temporary" files. On the
other hand, it means that theming information is now globally configured
(and applied) in `typst`, not split between `typst` and `make`. This way
theming becomes easier, too.